requirements.txt
scikit-image
scikit-learn
segment_anything
segmentation_models_pytorch
lightning
gym
python-box
torchsummary
tensorboard
albumentations
imagecorruptions